Update Performance Platform: Spotlight default
Spotlight is now the default, so we don't need all of this other
junk being deployed anymore.

publicapi is an app to proxy routes through to the backend via GOV.UK.
We stopped using it once the Performance Platform moved to its own
infrastructure, so the only thing that uses it is the Licensing
dashboards.