Simplify route primary key.
It's currently the tuple of incoming_path and route_type.  This was
needed to support an exact and prexif route at the same path.  This is
no longer a requirement, and this extra complexity has led to confusion.

This changes the logic to use just the incoming_path as the key for this
collection, and updates the API endpoints accordingly.

This doesn't update the unique index yet as some data cleanup will need
to be done first.