Read router_reload_urls from environment.
Instead of reading these from a yaml file that's overwritten on
deployment, read from an ENV variable.  This will allow us to remove
some logic from the deploy scripts.

This includes a fallback to continue to read from the yaml file if
present to ease transitioning to this new method.