Support HTTPS backends
Have the router-api consider backends using `https` as the URI scheme to
be valid URLs.

We're about to migrate most of the GOV.UK stack to a new provider, but
for accreditation reasons, Licensify will be continue to be hosted in
Skyscape.

Rather than reverse-proxy requests to the Licensify backend in Skyscape
further down the stack, have the router-api accept HTTPS URLs so that
the router can reverse-proxy requests to Licensify directly.

No changes to the router itself are necessary.