Add a rake task to modify backend_url
We need to modify backend_url in the router after the data sync
happens from production to staging and integration.

Initially this is just for the tariff app which is migrating to
the government PaaS, but eventually we should use this approach
to update all our application hostnames (this behaviour is
currently in env-sync-and-backup which makes queries directly
against a `mongo` shell which is not where that should live).