Fix race condition in RoutesController#update.
It was possible for two simultaneous requests to attempt to create a
route at the same path; find_or_initialize_by would fail to find the
route in both cases, and one would then succeed in saving the new route
but the second would fail.

The solution is to make this atomic, but we can't use the basic upsert
method because our unique key is a separate id, not the incoming_path.
Instead, we use the underlying Mongo find_and_modify method with
upsert:true, which allows us to specify which fields to search on to
find a document to update.

However, this does not call model validation or callbacks. We need to
validate the data before doing the call; unfortunately, this means we
need to remove the Mongoid uniqueness validator, as otherwise updates
would always conflict at this point. We still have a unique constraint
in the db itself though.

We also need to run callbacks explicitly. Currently there is only one,
on after_create, so we run this if the find_and_modify caused a create
only (there is no way of telling this directly, but that call returns
the document *before* the update, so we can tell it's a create if that
document is blank).