Fix race condition in RoutesController#update.
It was possible for two simultaneous requests to attempt to create a
route at the same path; find_or_initialize_by would fail to find the
route in both cases, and one would then succeed in saving the new route
but the second would fail.

The solution is to make this atomic, but we can't use the upsert
method because our unique key is a separate id, not the incoming_path.
Instead, we retry the update within the method a maximum of three
times.