Upgrade all backends from HTTP to HTTPS
Now that we have some confidence that the router is fine with HTTPS
backends (we've been using it across the internet for a while), we
should use HTTPS for our internal backends.

This adds a migration to upgrade all backends that are HTTP to be
HTTPS and ensures that the seed only creates HTTPS backends.