Allow forwarding of URL segments to campaigns
We are migrating our campaign formats from Publisher to the new Campaigns
platform.  Departments are using paid advertising to bring users to their
campaigns and need to keep the tracking data in the query string to work out
where people are coming from and therefore whether their advertising is value
for money.

This change permits *.campaign.gov.uk domain redirects to keep the query string
which contains that information.