Change route validation to use more specific error messages
Pull the complicated logic out so that each individual check is
connected to a error message. This makes changing the test suite
easier, but should also help users of the service.

This change also adds back in some of the more complicated
validations, e.g. that redirect targets cannot contain query
parameters if the segments mode is preserve.