Allow redirects to british-business-bank.co.uk
We have four external redirects that do not go to a .gov.uk subdomain. These
all [go to the same file][1] on [http://british-business-bank.co.uk][2].

These URLs are currently causing deploys of router-data to fail because router-api
currently only allows gov.uk subdomains for external redirects. We still need to
support these as we used to run applications supporting the Enterprise Finance
Guarantee before it was passed on to them.

Because of this, I'm extending the whitelist to include that domain too.  I've
checked the router database and we don't have redirects to any other non-gov.uk
domains, so hopefully this will remain a one off.

[1]: https://github.digital.cabinet-office.gov.uk/gds/router-data/blob/880166e47ecd79646db492043a05db994a7c8c2f/data/external-redirects.csv#L14-L17
[2]: http://british-business-bank.co.uk/wp-content/uploads/2015/02/BBB-011114-44-EFG-business-sectors-and-purposes.pdf