Add set_router_reload_urls_from_file method
Traditionally we have been able to pass in the router nodes to
router-api using an environment variable as we hardcoded the hostnames
that the instance needed to connect to.

In a dynamic environment, such as AWS, it is no longer possible to do
this as we do not choose hostnames assigned to instances.

Instead, we can create a script that will write out the current router
nodes to a file. If any nodes get rebuilt or change, the file will be
updated. router-api can then read this file in and connect to the
relevant nodes to reload routes.