Correct backend_id for statistics pages
This is a quick fix for a statistics page with an incorrect
backend_id. The page is rendered by whitehall-frontend, but
should be rendered by government-frontend, like other stats
announcements.

There are ~1,000 other stats announcements affected, so we'll
probably need to run this for the others in future once we have
fixed the underlying bug.

Zendesk ticket: https://govuk.zendesk.com/agent/tickets/3773523