Add a temp rake task to redirect legacy signups
This commit adds a temporary rake task to create redirects for
specialist sector email signup URL's.

The current behaviour is for these specialist sector email-signup URL's
to be sent to collections which then redirects to email-alert-frontend.

Now the redirect happens soley in the router and points to the parent
page rendered by collections.

The list below may get outdated between this being reviewed and deployed
but before deploying I'll pull a fresh list from collections-publisher.

Depends on:
https://github.com/alphagov/collections-publisher/pull/1102

---

Trello:
https://trello.com/c/Fi2Z2zDy/399-redirect-and-remove-legacy-email-signup-routes-for-collections