Add backend:add_subdomains rake task
This will add a subdomain_name to each Backend that currently
has routes which are actually handled by a backend (not gone
or redirect routes). The subdomain_name should match the name
at the beginning of the FQDN for that backend. E.g. if the app
has a domain name of frontend.integration.govuk-internal.digital
we should provide Router with the ROUTE_DOMAIN env var of
integration.govuk-internal.digital and set the subdomain_name
for the frontend backend in Router API to 'frontend'.