Refactor redirect validation to remove hostname validation
Since this app consumes data originating from the Publishing API it
seems unnecessary for this to replicate the URI validation that is done
there, in particular with regard to host validation - as there becomes a
degree of repetition if any changes are made there.

I've therefore simplified the validation of redirects to be much more
validity focused rather than organisational preferences. These
preferences should occur earlier in the stack (Publishing API).

The motivation for this change was a need to add nhs.uk to the allowed
hosts that could be redirected to and it seemed frustrating to do that
in here and the Publishing API.