Specify empty Router URL list if none provided
This commit changes the behaviour of the `urls` method to use an empty list if no router node endpoints can be found from either `ROUTER_NODES` or `ROUTER_NODES_FILE`. This change is part of the work to have Router update itself and no longer rely on Router API to manually call Routers' `/reload` endpoints.